% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AUC_Glucose.R
\name{calculate_AUC_Glucose}
\alias{calculate_AUC_Glucose}
\title{Calculate AUC for Glucose}
\usage{
calculate_AUC_Glucose(time, glucose)
}
\arguments{
\item{time}{Numeric vector of time points.}

\item{glucose}{Numeric vector of glucose levels (mg/dL) corresponding to the time points.}
}
\value{
A numeric value representing the AUC.
}
\description{
This function calculates the area under the curve (AUC) for glucose levels using the trapezoidal rule.
}
\examples{
calculate_AUC_Glucose(c(0, 30, 60, 120), c(90, 140, 120, 100))
}
