% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BMI.R
\name{calculate_BMI}
\alias{calculate_BMI}
\title{Calculate BMI}
\usage{
calculate_BMI(weight, height)
}
\arguments{
\item{weight}{Numeric, weight in kilograms.}

\item{height}{Numeric, height in meters.}
}
\value{
A numeric value representing the BMI.
}
\description{
This function calculates the Body Mass Index (BMI).
}
\examples{
calculate_BMI(70, 1.75)
}
