% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BMR.R
\name{calculate_BMR}
\alias{calculate_BMR}
\title{Calculate Basal Metabolic Rate (BMR)}
\usage{
calculate_BMR(weight, height, age, sex)
}
\arguments{
\item{weight}{Numeric, weight in kilograms.}

\item{height}{Numeric, height in centimeters.}

\item{age}{Numeric, age in years.}

\item{sex}{Character, "male" or "female".}
}
\value{
A numeric value representing the BMR.
}
\description{
This function calculates the Basal Metabolic Rate (BMR) using the Harris-Benedict formula.
}
\examples{
calculate_BMR(70, 175, 25, "male")
}
