% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_BodyFatPercentage.R
\name{calculate_BodyFatPercentage}
\alias{calculate_BodyFatPercentage}
\title{Calculate Body Fat Percentage}
\usage{
calculate_BodyFatPercentage(bmi, age, sex)
}
\arguments{
\item{bmi}{Numeric, Body Mass Index.}

\item{age}{Numeric, age in years.}

\item{sex}{Character, "male" or "female".}
}
\value{
A numeric value representing the body fat percentage.
}
\description{
This function calculates the body fat percentage using the Deurenberg formula.
}
\examples{
calculate_BodyFatPercentage(22, 25, "male")
}
