% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CardiacRiskRatio.R
\name{calculate_CardiacRiskRatio}
\alias{calculate_CardiacRiskRatio}
\title{Calculate Cardiac Risk Ratio}
\usage{
calculate_CardiacRiskRatio(total_cholesterol, hdl)
}
\arguments{
\item{total_cholesterol}{Numeric, total cholesterol level (mg/dL).}

\item{hdl}{Numeric, HDL cholesterol level (mg/dL).}
}
\value{
A numeric value representing the cardiac risk ratio.
}
\description{
This function calculates the cardiac risk ratio (Total Cholesterol / HDL).
}
\examples{
calculate_CardiacRiskRatio(200, 50)
}
