% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_GlucoseVariability.R
\name{calculate_GlucoseVariability}
\alias{calculate_GlucoseVariability}
\title{Calculate Glucose Variability}
\usage{
calculate_GlucoseVariability(glucose_vector)
}
\arguments{
\item{glucose_vector}{Numeric vector of glucose levels (mg/dL).}
}
\value{
A numeric value representing the standard deviation of glucose.
}
\description{
This function calculates the standard deviation of glucose levels.
}
\examples{
calculate_GlucoseVariability(c(90, 95, 100, 85, 110))
}
