% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_HDL_LDL_Ratio.R
\name{calculate_HDL_LDL_Ratio}
\alias{calculate_HDL_LDL_Ratio}
\title{Calculate HDL to LDL Ratio}
\usage{
calculate_HDL_LDL_Ratio(hdl, ldl)
}
\arguments{
\item{hdl}{Numeric, HDL cholesterol level (mg/dL).}

\item{ldl}{Numeric, LDL cholesterol level (mg/dL).}
}
\value{
A numeric value representing the HDL to LDL ratio.
}
\description{
This function calculates the ratio of HDL to LDL cholesterol.
}
\examples{
calculate_HDL_LDL_Ratio(50, 100)
}
