% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_HOMA.R
\name{calculate_HOMA}
\alias{calculate_HOMA}
\title{Calculate HOMA-IR Index}
\usage{
calculate_HOMA(glucose, insulin)
}
\arguments{
\item{glucose}{Numeric, fasting glucose level (mg/dL).}

\item{insulin}{Numeric, fasting insulin level (uU/mL).}
}
\value{
A numeric value representing the HOMA-IR index.
}
\description{
This function calculates the HOMA-IR index (Homeostatic Model Assessment of Insulin Resistance).
It is used to estimate insulin resistance.
}
\examples{
calculate_HOMA(90, 15)
}
