% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_HOMA_beta.R
\name{calculate_HOMA_beta}
\alias{calculate_HOMA_beta}
\title{Calculate HOMA-Beta}
\usage{
calculate_HOMA_beta(glucose, insulin)
}
\arguments{
\item{glucose}{Numeric, fasting glucose level (mg/dL).}

\item{insulin}{Numeric, fasting insulin level (uU/mL).}
}
\value{
A numeric value representing the HOMA-Beta index.
}
\description{
This function calculates the HOMA-Beta index to estimate beta-cell function.
}
\examples{
calculate_HOMA_beta(90, 15)
}
