% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_InsulinogenicIndex.R
\name{calculate_InsulinogenicIndex}
\alias{calculate_InsulinogenicIndex}
\title{Calculate Insulinogenic Index}
\usage{
calculate_InsulinogenicIndex(
  insulin_30min,
  insulin_fasting,
  glucose_30min,
  glucose_fasting
)
}
\arguments{
\item{insulin_30min}{Numeric, insulin level at 30 minutes (uU/mL).}

\item{insulin_fasting}{Numeric, fasting insulin level (uU/mL).}

\item{glucose_30min}{Numeric, glucose level at 30 minutes (mg/dL).}

\item{glucose_fasting}{Numeric, fasting glucose level (mg/dL).}
}
\value{
A numeric value representing the insulinogenic index.
}
\description{
This function calculates the insulinogenic index, measuring early-phase insulin response.
}
\examples{
calculate_InsulinogenicIndex(50, 15, 130, 90)
}
