% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LDL.R
\name{calculate_LDL}
\alias{calculate_LDL}
\title{Calculate LDL Cholesterol}
\usage{
calculate_LDL(total_cholesterol, hdl, triglycerides)
}
\arguments{
\item{total_cholesterol}{Numeric, total cholesterol level (mg/dL).}

\item{hdl}{Numeric, HDL cholesterol level (mg/dL).}

\item{triglycerides}{Numeric, triglyceride level (mg/dL).}
}
\value{
A numeric value representing LDL cholesterol (mg/dL).
}
\description{
This function calculates LDL cholesterol using the Friedewald formula.
}
\examples{
calculate_LDL(200, 50, 150)
}
