% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_LeanBodyMass.R
\name{calculate_LeanBodyMass}
\alias{calculate_LeanBodyMass}
\title{Calculate Lean Body Mass (LBM)}
\usage{
calculate_LeanBodyMass(weight, height, sex)
}
\arguments{
\item{weight}{Numeric, weight in kilograms.}

\item{height}{Numeric, height in centimeters.}

\item{sex}{Character, "male" or "female".}
}
\value{
A numeric value representing the Lean Body Mass (LBM).
}
\description{
This function calculates the Lean Body Mass (LBM) using Boer’s formula.
}
\examples{
calculate_LeanBodyMass(70, 175, "male")
}
