% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MAP.R
\name{calculate_MAP}
\alias{calculate_MAP}
\title{Calculate Mean Arterial Pressure (MAP)}
\usage{
calculate_MAP(sbp, dbp)
}
\arguments{
\item{sbp}{Numeric, systolic blood pressure (mmHg).}

\item{dbp}{Numeric, diastolic blood pressure (mmHg).}
}
\value{
A numeric value representing the Mean Arterial Pressure (MAP).
}
\description{
This function calculates the Mean Arterial Pressure (MAP).
}
\examples{
calculate_MAP(120, 80)
}
