% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_MatsudaIndex.R
\name{calculate_MatsudaIndex}
\alias{calculate_MatsudaIndex}
\title{Calculate Matsuda Index}
\usage{
calculate_MatsudaIndex(
  glucose_fasting,
  glucose_120min,
  insulin_fasting,
  insulin_120min
)
}
\arguments{
\item{glucose_fasting}{Numeric, fasting glucose level (mg/dL).}

\item{glucose_120min}{Numeric, glucose level at 120 minutes (mg/dL).}

\item{insulin_fasting}{Numeric, fasting insulin level (uU/mL).}

\item{insulin_120min}{Numeric, insulin level at 120 minutes (uU/mL).}
}
\value{
A numeric value representing the Matsuda Index.
}
\description{
This function calculates the Matsuda Index to estimate insulin sensitivity.
}
\examples{
calculate_MatsudaIndex(90, 140, 15, 50)
}
