% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_NonHDL.R
\name{calculate_NonHDL}
\alias{calculate_NonHDL}
\title{Calculate Non-HDL Cholesterol}
\usage{
calculate_NonHDL(total_cholesterol, hdl)
}
\arguments{
\item{total_cholesterol}{Numeric, total cholesterol level (mg/dL).}

\item{hdl}{Numeric, HDL cholesterol level (mg/dL).}
}
\value{
A numeric value representing Non-HDL cholesterol (mg/dL).
}
\description{
This function calculates Non-HDL cholesterol.
}
\examples{
calculate_NonHDL(200, 50)
}
