% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PostprandialDelta.R
\name{calculate_PostprandialDelta}
\alias{calculate_PostprandialDelta}
\title{Calculate Postprandial Delta}
\usage{
calculate_PostprandialDelta(glucose_fasting, glucose_postprandial)
}
\arguments{
\item{glucose_fasting}{Numeric, fasting glucose level (mg/dL).}

\item{glucose_postprandial}{Numeric, postprandial glucose level (mg/dL).}
}
\value{
A numeric value representing the difference in glucose levels.
}
\description{
This function calculates the difference between fasting and postprandial glucose levels.
}
\examples{
calculate_PostprandialDelta(90, 140)
}
