% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_PulsePressure.R
\name{calculate_PulsePressure}
\alias{calculate_PulsePressure}
\title{Calculate Pulse Pressure}
\usage{
calculate_PulsePressure(sbp, dbp)
}
\arguments{
\item{sbp}{Numeric, systolic blood pressure (mmHg).}

\item{dbp}{Numeric, diastolic blood pressure (mmHg).}
}
\value{
A numeric value representing the pulse pressure (mmHg).
}
\description{
This function calculates the pulse pressure (SBP - DBP).
}
\examples{
calculate_PulsePressure(120, 80)
}
