% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_QUICKI.R
\name{calculate_QUICKI}
\alias{calculate_QUICKI}
\title{Calculate QUICKI Index}
\usage{
calculate_QUICKI(glucose, insulin)
}
\arguments{
\item{glucose}{Numeric, fasting glucose level (mg/dL).}

\item{insulin}{Numeric, fasting insulin level (uU/mL).}
}
\value{
A numeric value representing the QUICKI index.
}
\description{
This function calculates the QUICKI index (Quantitative Insulin Sensitivity Check Index).
It is used as an estimate of insulin sensitivity.
}
\examples{
calculate_QUICKI(90, 15)
}
