% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RMR.R
\name{calculate_RMR}
\alias{calculate_RMR}
\title{Calculate Resting Metabolic Rate (RMR)}
\usage{
calculate_RMR(weight, height, age, sex)
}
\arguments{
\item{weight}{Numeric, weight in kilograms.}

\item{height}{Numeric, height in centimeters.}

\item{age}{Numeric, age in years.}

\item{sex}{Character, "male" or "female".}
}
\value{
A numeric value representing the RMR.
}
\description{
This function calculates the Resting Metabolic Rate (RMR).
}
\examples{
calculate_RMR(70, 175, 25, "male")
}
