% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_TG_HDL_Ratio.R
\name{calculate_TG_HDL_Ratio}
\alias{calculate_TG_HDL_Ratio}
\title{Calculate Triglyceride to HDL Ratio}
\usage{
calculate_TG_HDL_Ratio(triglycerides, hdl)
}
\arguments{
\item{triglycerides}{Numeric, triglyceride level (mg/dL).}

\item{hdl}{Numeric, HDL cholesterol level (mg/dL).}
}
\value{
A numeric value representing the triglyceride to HDL ratio.
}
\description{
This function calculates the ratio of triglycerides to HDL cholesterol.
}
\examples{
calculate_TG_HDL_Ratio(150, 50)
}
