% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_WaistToHeight.R
\name{calculate_WaistToHeight}
\alias{calculate_WaistToHeight}
\title{Calculate Waist-to-Height Ratio (WHtR)}
\usage{
calculate_WaistToHeight(waist, height)
}
\arguments{
\item{waist}{Numeric, waist circumference in centimeters.}

\item{height}{Numeric, height in centimeters.}
}
\value{
A numeric value representing the WHtR.
}
\description{
This function calculates the Waist-to-Height Ratio.
}
\examples{
calculate_WaistToHeight(80, 175)
}
