% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_WHR.R
\name{calculate_WHR}
\alias{calculate_WHR}
\title{Calculate Waist-to-Hip Ratio (WHR)}
\usage{
calculate_WHR(waist, hip)
}
\arguments{
\item{waist}{Numeric, waist circumference in centimeters.}

\item{hip}{Numeric, hip circumference in centimeters.}
}
\value{
A numeric value representing the WHR.
}
\description{
This function calculates the Waist-to-Hip Ratio.
}
\examples{
calculate_WHR(80, 100)
}
